requirements.txt
discord.py==1.3.4
http3==0.6.7
motor==2.1.0
pytz==2020.1
pyyaml==5.3.1
redis
https://github.com/franagustin/bori-models/archive/master.zip
